requirements.txt
Flask
pandas
scikit-surprise
flask-cors
